+++ src/executions/path.c
@@ -6,11 +6,11 @@
 /*   By: jaeshin <jaeshin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 15:08:15 by jaeshin           #+#    #+#             */
-/*   Updated: 2023/09/04 15:08:16 by jaeshin          ###   ########.fr       */
+/*   Updated: 2023/09/04 21:10:08 by jaeshin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
+#include "../../inc/minishell.h"
 
 // free the temporary containers after usages
 void	free_container(char **temp)
@@ -63,7 +63,7 @@ char	*get_path(char *cmd, char **envp)
 
 	i = 0;
 	all_path = ft_split(find_path("PATH", envp), ':');
-	split_cmd = ft_split(cmd, ' ');
+	split_cmd = space_quotes_split(cmd);
 	while (all_path[i])
 	{
 		one_path = ft_strjoin(all_path[i], "/");
@@ -80,29 +80,4 @@ char	*get_path(char *cmd, char **envp)
 	free_container(all_path);
 	free_container(split_cmd);
 	return (cmd);
-}
-
-// tries to run the command, if it doesn't work, returns the error message.
-void	execute(char *cmd, t_cmd *cmd_args)
-{
-	char	**split_cmd;
-	int		result;
-
-	split_cmd = ft_split(cmd, ' ');
-	result = 0;
-	if (ft_strncmp(split_cmd[0], "cd", sizeof(split_cmd[0])) == 0)
-		result = handle_cd(split_cmd, cmd_args);
-	// PWD edit can be here
-	else if (ft_strncmp(split_cmd[0], "pwd", sizeof(split_cmd[0])) == 0)
-		printf("%s\n", cmd_args->abs_path);
-	else if (ft_strncmp(split_cmd[0], "unset", sizeof(split_cmd[0])) == 0
-		&& ft_strncmp(split_cmd[1], "PATH", sizeof(split_cmd[0])) == 0)
-		cmd_args->envp = NULL;
-	else
-	{
-		printf("%s\n", cmd_args->envp[0]);
-		result = handle_else(split_cmd, cmd_args);
-	}
-	result_error(result, split_cmd);
-	exit(0);
 }
